+++ src/libwayland-shim.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include "libwayland-shim.h"
 #include "layer-surface.h"
-#include "wayland-utils.h"
+#include "stolen-from-libwayland.h"
 
 struct wl_proxy* (*libwayland_shim_real_wl_proxy_marshal_array_flags)(
     struct wl_proxy* proxy,
@@ -48,84 +48,6 @@ static void libwayland_shim_init() {
 #undef INIT_SYM
 
     //dlclose(handle);
-}
-
-// From wayland-private.h in libwayland
-#define WL_CLOSURE_MAX_ARGS 20
-
-// From wayland-private.h in libwayland
-struct argument_details {
-	char type;
-	int nullable;
-};
-
-// From connection.c in libwayland
-static const char *
-get_next_argument(const char *signature, struct argument_details *details)
-{
-	details->nullable = 0;
-	for(; *signature; ++signature) {
-		switch(*signature) {
-		case 'i':
-		case 'u':
-		case 'f':
-		case 's':
-		case 'o':
-		case 'n':
-		case 'a':
-		case 'h':
-			details->type = *signature;
-			return signature + 1;
-		case '?':
-			details->nullable = 1;
-		}
-	}
-	details->type = '\0';
-	return signature;
-}
-
-// From connection.c in libwayland
-static void
-wl_argument_from_va_list(const char *signature, union wl_argument *args,
-			 int count, va_list ap)
-{
-	int i;
-	const char *sig_iter;
-	struct argument_details arg;
-
-	sig_iter = signature;
-	for (i = 0; i < count; i++) {
-		sig_iter = get_next_argument(sig_iter, &arg);
-
-		switch(arg.type) {
-		case 'i':
-			args[i].i = va_arg(ap, int32_t);
-			break;
-		case 'u':
-			args[i].u = va_arg(ap, uint32_t);
-			break;
-		case 'f':
-			args[i].f = va_arg(ap, wl_fixed_t);
-			break;
-		case 's':
-			args[i].s = va_arg(ap, const char *);
-			break;
-		case 'o':
-			args[i].o = va_arg(ap, struct wl_object *);
-			break;
-		case 'n':
-			args[i].o = va_arg(ap, struct wl_object *);
-			break;
-		case 'a':
-			args[i].a = va_arg(ap, struct wl_array *);
-			break;
-		case 'h':
-			args[i].h = va_arg(ap, int32_t);
-			break;
-		case '\0':
-			return;
-		}
-	}
 }
 
 struct wrapped_proxy {
@@ -274,7 +196,8 @@ struct wl_proxy* wl_proxy_marshal_array_flags(
         return result;
     } else {
         struct wl_proxy* ret_proxy = NULL;
-        if (layer_surface_handle_request(proxy, opcode, interface, version, flags, args, &ret_proxy)) {
+        const char* type_name = proxy->object.interface->name;
+        if (layer_surface_handle_request(type_name, proxy, opcode, interface, version, flags, args, &ret_proxy)) {
             // The behavior of the request has been overridden
             return ret_proxy;
         } else if (args_contains_client_facing_proxy(proxy, opcode, interface, args)) {
@@ -310,109 +233,10 @@ int wl_proxy_add_dispatcher(
     return libwayland_shim_real_wl_proxy_add_dispatcher(proxy, dispatcher_func, dispatcher_data, data);
 }
 
-// The following functions may be unnecessary, as they are unmodified from libwayland, however it is possible that some
-// builds of libwayland could inline the call they make to wl_proxy_marshal_array_flags(), which means we need to
-// override all functions that call it.
-
-// Overrides the function in wayland-client.c in libwayland
-struct wl_proxy *
-wl_proxy_marshal_flags(struct wl_proxy *proxy, uint32_t opcode,
-		       const struct wl_interface *interface,
-		       uint32_t version,
-		       uint32_t flags, ...)
-{
-    union wl_argument args[WL_CLOSURE_MAX_ARGS];
-	va_list ap;
-
-	va_start(ap, flags);
-	wl_argument_from_va_list(proxy->object.interface->methods[opcode].signature,
-				 args, WL_CLOSURE_MAX_ARGS, ap);
-	va_end(ap);
-
-	return wl_proxy_marshal_array_flags(proxy, opcode, interface, version, flags, args);
+void const* libwayland_shim_proxy_get_implementation(struct wl_proxy* proxy) {
+    return proxy->object.implementation;
 }
 
-// Overrides the function in wayland-client.c in libwayland
-void
-wl_proxy_marshal(struct wl_proxy *proxy, uint32_t opcode, ...)
-{
-	union wl_argument args[WL_CLOSURE_MAX_ARGS];
-	va_list ap;
-
-	va_start(ap, opcode);
-	wl_argument_from_va_list(proxy->object.interface->methods[opcode].signature,
-				 args, WL_CLOSURE_MAX_ARGS, ap);
-	va_end(ap);
-
-	wl_proxy_marshal_array_constructor(proxy, opcode, args, NULL);
-}
-
-// Overrides the function in wayland-client.c in libwayland
-void
-wl_proxy_marshal_array(struct wl_proxy *proxy, uint32_t opcode,
-		       union wl_argument *args)
-{
-	wl_proxy_marshal_array_constructor(proxy, opcode, args, NULL);
-}
-
-// Overrides the function in wayland-client.c in libwayland
-struct wl_proxy *
-wl_proxy_marshal_constructor(struct wl_proxy *proxy,
-			     uint32_t opcode,
-			     const struct wl_interface *interface,
-			     ...)
-{
-	union wl_argument args[WL_CLOSURE_MAX_ARGS];
-	va_list ap;
-
-	va_start(ap, interface);
-	wl_argument_from_va_list(proxy->object.interface->methods[opcode].signature,
-				 args, WL_CLOSURE_MAX_ARGS, ap);
-	va_end(ap);
-
-	return wl_proxy_marshal_array_constructor(proxy, opcode,
-						  args, interface);
-}
-
-// Overrides the function in wayland-client.c in libwayland
-struct wl_proxy *
-wl_proxy_marshal_constructor_versioned(struct wl_proxy *proxy,
-				       uint32_t opcode,
-				       const struct wl_interface *interface,
-				       uint32_t version,
-				       ...)
-{
-	union wl_argument args[WL_CLOSURE_MAX_ARGS];
-	va_list ap;
-
-	va_start(ap, version);
-	wl_argument_from_va_list(proxy->object.interface->methods[opcode].signature,
-				 args, WL_CLOSURE_MAX_ARGS, ap);
-	va_end(ap);
-
-	return wl_proxy_marshal_array_constructor_versioned(proxy, opcode,
-							    args, interface,
-							    version);
-}
-
-// Overrides the function in wayland-client.c in libwayland
-struct wl_proxy *
-wl_proxy_marshal_array_constructor(struct wl_proxy *proxy,
-				   uint32_t opcode, union wl_argument *args,
-				   const struct wl_interface *interface)
-{
-	return wl_proxy_marshal_array_constructor_versioned(proxy, opcode,
-							    args, interface,
-							    proxy->version);
-}
-
-// Overrides the function in wayland-client.c in libwayland
-struct wl_proxy *
-wl_proxy_marshal_array_constructor_versioned(struct wl_proxy *proxy,
-					     uint32_t opcode,
-					     union wl_argument *args,
-					     const struct wl_interface *interface,
-					     uint32_t version)
-{
-	return wl_proxy_marshal_array_flags(proxy, opcode, interface, version, 0, args);
+void* libwayland_shim_proxy_get_user_data(struct wl_proxy* proxy) {
+    return proxy->user_data;
 }
